requirements.txt
# === Web Framework ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# === Document Processing ===
docling>=2.15.0
easyocr>=1.7.1

# === VLM Integration ===
openai>=1.40.0
httpx>=0.27.0

# === Text Processing & Chunking ===
langchain-text-splitters>=0.2.0

# === Vector Database ===
chromadb>=0.5.0
sentence-transformers>=3.0.0

# === Configuration ===
pydantic-settings>=2.4.0
python-dotenv>=1.0.0
pyyaml>=6.0

# === Image Processing ===
Pillow>=10.0.0
pdf2image>=1.17.0

# === Async Support ===
aiofiles>=24.1.0

# === Testing ===
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
httpx>=0.27.0

# === Utilities ===
loguru>=0.7.0
